+++ src/HarkestDungeon/Screens/AltarRecollectionScreen.cs
@@ -5,7 +5,6 @@ using DD2A11y.Elements;
 using DD2A11y.Game;
 using HarmonyLib;
 using S = DD2A11y.Core.Strings.Strings;
-using UnityEngine;
 using UnityEngine.UI;
 
 namespace DD2A11y.Screens {
@@ -29,8 +28,7 @@ namespace DD2A11y.Screens {
 
         public override string Name {
             get {
-                var anchor = FindChild(_panel != null ? _panel.transform : null, "exit_anchor");
-                string title = anchor == null ? null : UiText.FirstLabel(anchor.gameObject);
+                string title = UiText.ChildLabel(_panel != null ? _panel.gameObject : null, "exit_anchor");
                 return string.IsNullOrEmpty(title) ? S.ScreenGeneric : title;
             }
         }
@@ -81,18 +79,6 @@ namespace DD2A11y.Screens {
                 }
             }
             return _root;
-        }
-
-        private static Transform FindChild(Transform root, string name) {
-            if (root == null) {
-                return null;
-            }
-            foreach (var child in root.GetComponentsInChildren<Transform>(includeInactive: false)) {
-                if (child.name == name) {
-                    return child;
-                }
-            }
-            return null;
         }
     }
 }